SW-1327 Move the creation of the outline path from preview to on update
The outline path is being created for every QuickText string regardless of the option being disabled. For this to work we were hiding the path if the option is disabled via setting the d attribute to an empty string. This created an issue when the mrb file is created via a QuickText SVG where the path was considered as a valid one. This also creates a bug with the job time estimation as it shows a path with NaN values as an estimate.