SW-1326 Set stroke in the text element to none
This stroke is used to detect the availability of an outline. The outline will be handled in the .qt_outline element. This should be removed at the last point before added to the working area otherwise it will be lasered as a thick stroke included with the filling whenever filling and outline are checked together.