SW-1326 the outline and the engraving is not matching the quicktext preview (#1541)
* SW-1326 Modify fill attribute and fill-opacity style

These attribute and styling should always be set to white and zero so that the outline path always have transparent filling. The filling of the quicktext is handled by another text element.

* SW-1326 Set stroke in the text element to none

This stroke is used to detect the availability of an outline. The outline will be handled in the .qt_outline element. This should be removed at the last point before added to the working area otherwise it will be lasered as a thick stroke included with the filling whenever filling and outline are checked together.