SW-2835 Fix the bug in a different manner
This bug is caused by the method to convert elements into strings when they are not part of the DOM. The bug is resolved initially by handling certain characters to avoid this from happening, but it was a patchy and incomplete solution as it didn't handle the styling of these elements as well. This solution appends the elements to the DOM, convert them to strings and them removes them.